#1721 use a single selection list for getting case index data
